Tools: Add script for fixing priviledges for OpenOCD and serial console on Linux
This script will essentially add the Linux user to 'plugdev' and 'dialout' groups
and also add udev rules for OpenOCD. User will need to log out and log-in once after
this step. After this, sudo is not required for using OpenOCD or serial console.

This script is mandatory for Eclipse and arm-none-eabi-gdb users on Linux

Signed-off-by: Rahul Masurkar <rahulgm@marvell.com>